[NI-CompactDAQ] Input Stream | App Deme

+++ blocks/DATA/GENERATION/SIMULATIONS/VECTOR/VECTOR.py
@@ -1,5 +1,6 @@
 from flojoy import flojoy, Vector
 from numpy import array
+import numpy as np
 from typing import Literal
 
 
@@ -49,6 +50,6 @@ def VECTOR(
                     f"all elements of the vector must be numeric: {element}"
                 )
 
-        return Vector(v=array(all_numeric))
+        return Vector(v=array(all_numeric, dtype=np.int16))
 
     raise ValueError("all elements of the vector must be in boolean or integer type")
